requirements.txt
Pillow>=9.0.0
watchdog>=2.1.0